+++ superset/commands/chart/delete.py
@@ -45,7 +45,8 @@ class DeleteChartCommand(BaseCommand):
     @transaction(on_error=partial(on_error, reraise=ChartDeleteFailedError))
     def run(self) -> None:
         self.validate()
-        assert self._models
+        if not self._models:
+            raise ChartNotFoundError()
         ChartDAO.delete(self._models)
 
     def validate(self) -> None:
